# FIX: more performance improvement for PostAlert job