--- a/app/models/topic_user.rb
+++ b/app/models/topic_user.rb
@@ -582,6 +582,7 @@ end
 #
 # Indexes
 #
-#  index_topic_users_on_topic_id_and_user_id  (topic_id,user_id) UNIQUE
-#  index_topic_users_on_user_id_and_topic_id  (user_id,topic_id) UNIQUE
+#  index_topic_users_on_topic_id_and_notification_level  (topic_id,notification_level)
+#  index_topic_users_on_topic_id_and_user_id             (topic_id,user_id) UNIQUE
+#  index_topic_users_on_user_id_and_topic_id             (user_id,topic_id) UNIQUE
 #