--- a/app/models/user_option.rb
+++ b/app/models/user_option.rb
@@ -293,6 +293,7 @@ end
 #
 # Indexes
 #
-#  index_user_options_on_user_id                       (user_id) UNIQUE
-#  index_user_options_on_user_id_and_default_calendar  (user_id,default_calendar)
+#  index_user_options_on_user_id                                    (user_id) UNIQUE
+#  index_user_options_on_user_id_and_default_calendar               (user_id,default_calendar)
+#  index_user_options_on_user_id_and_watched_precedence_over_muted  (user_id,watched_precedence_over_muted)
 #