--- a/db/migrate/20230712011946_add_topic_id_notification_level_index_to_topic_users.rb
+++ b/db/migrate/20230712011946_add_topic_id_notification_level_index_to_topic_users.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+#
+class AddTopicIdNotificationLevelIndexToTopicUsers < ActiveRecord::Migration[7.0]
+  def change
+    add_index :topic_users, %i[topic_id notification_level]
+  end
+end