--- a/db/migrate/20230712013248_add_category_id_notification_level_index_to_category_users.rb
+++ b/db/migrate/20230712013248_add_category_id_notification_level_index_to_category_users.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddCategoryIdNotificationLevelIndexToCategoryUsers < ActiveRecord::Migration[7.0]
+  def change
+    add_index :category_users, %i[category_id notification_level]
+  end
+end